### 2017-09-24 #409 @biorelated
@nelkibs Would you be happy to review this work for the JOSS journal? I am happy to assist you with the review process. Thanks!
