### 2017-09-29 #409 @nelkibs
Hi George,

Been off radar since last week. Technological problem. Just seen this. Si
we talk about it on Monday?

Nelson
