### 2017-10-30 #445 @whedon
[REVIEW]: CytobankAPIstats: Computes Signaling and Population Statistics for Cytometry Data on Cytobank using 'CytobankAPI'
**Submitting author:** @athrom (<a href="http://orcid.org/0000-0001-6689-7904">Allison Throm</a>)
**Repository:** <a href="https://github.com/athrom/CytobankAPIstats" target ="_blank">https://github.com/athrom/CytobankAPIstats</a>
**Version:** v1.0
