### 2018-02-12 #445 @nelkibs
@biorelated and @arfon Sorry for the very slow response. I have been off radar for a while. The checkbox button are inactive on my end? Is there a setting I need to change?
