### 2018-02-12 #445 @arfon
@nelkibs - you should have permissions to edit the checklist - GitHub requires a pretty modern browser - which are you using?
