### 2018-03-19 #445 @biorelated
@nelkibs raised some issues in regard to this submission.  @athrom as not commented on the issues.
@athrom could kindly comment on these issues?
