### 2018-03-20 #445 @athrom
I have addressed the issues raised by adding LICENSE and INSTALLATION INSTRUCTIONS documents to github.
