### 2018-11-03 #445 @arfon
@pjotrp or @mgymrek - would either of you be able to take over as the editor here?
