### 2018-11-13 #445 @arfon
Thanks @nelkibs.

> You can however add a little more information on community guidelines (e.g how a user can seek help or extend the package etc).

@athrom - do you think you could do this ☝️
