### 2018-11-14 0eb18e6
update readme to add more community guidelines
