+++ README.md
@@ -2,3 +2,7 @@
 Contains code and documentation for "CytobankAPIstats" library available on CRAN
 R folder contains code for all functions for packages
 man foolder contains documentation for help files
+
+For help using the package, please see the CRAN repository documentation: https://cran.r-project.org/web/packages/CytobankAPIstats/CytobankAPIstats.pdf.
+
+If interested in extending the package, please cite the original CRAN package: https://cran.r-project.org/web/packages/CytobankAPIstats/index.html
