### 2018-11-14 #445 @athrom
Hi @nelkibs @arfon
I added the followinglines to the README file:

> For help using the package, please see the CRAN repository documentation: https://cran.r-project.org/web/packages/CytobankAPIstats/CytobankAPIstats.pdf.
>
> If interested in extending the package, please cite the original CRAN package: https://cran.r-project.org/web/packages/CytobankAPIstats/index.html

Please let me know if there is anything else I need to address/add to this declaration.
