### 2018-11-14 #445 @george-githinji
Hi @nelkibs please confirm if this is satisfactory.
