### 2018-12-28 #445 @pjotrp
@athrom - sorry for the lengthy process. I am here to help now. Your publication is almost ready. Can you fix the PDF (regenerate with above command). Also to finalize your submission and accept your paper in JOSS, we need two things. First, can you confirm that all references in your bibliography have a DOI (if one exists).

Second, we need you to deposit a copy of your software repository (including any revisions made during the JOSS review process) with a data-archiving service. To do so:

 3. Post a comment here with the DOI for the release.
