+++ paper.md
@@ -1,5 +1,8 @@
 ---
   title: "CytobankAPIstats: Computes Signaling and Population Statistics for Cytometry Data on Cytobank using 'CytobankAPI'"
+  tags:
+   - R
+   - mass cytometry
   authors:
    - name: Allison A. Throm
      affiliation: 1,2
