### 2019-03-05 #445 @kyleniemeyer
Hi @athrom @pjotrp I'm going to send this back to you for some more work, because the paper itself does not quite meet [our expectations](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). Right now the article itself is extremely brief and doesn't contain much substance; while we don't expect a full article here, it should be 250–1000 words (currently the main article is <150, which put together is a long paragraph). To me, these items need to be improved:

- A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience
- A clear statement of need that illustrates the purpose of the software

Further, I think the article should be improved with more description of how the software works, some examples, and use cases.

Also, the first reference has some issues in the way the DOI is being presented, I think.
