### 2019-03-05 #445 @pjotrp
@kyleniemeyer I think the paper is fine. It is not meant for a non-specialist audience and I understand it. The purpose is stated too. It is an impressively short paper - which is usually hard to do.

Yes, the first DOI may be wrong, but the system did not catch that.
