### 2019-03-05 #445 @arfon
> @kyleniemeyer I think the paper is fine. It is not meant for a non-specialist audience and I understand it. The purpose is stated too. It is an impressively short paper - which is usually hard to do.

This is incorrect. Quoting directly from https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain

> JOSS welcomes submissions from broadly diverse research areas. For this reason, we request that authors include in the paper some sentences that would explain the software functionality and domain of use to a non-specialist reader. Your submission should probably be somewhere between 250-1000 words.
