### 2019-03-05 #445 @kyleniemeyer
@pjotrp Furthermore, the @openjournals/joss-eics have agreed that it's our responsibility to give a final pass over a submission, such that we can give feedback to the author(s), reviewers, and editor (if needed) on the expectations we have agreed on.

If you have issues with that, we can take this offline (... online?) to discuss over email, rather than in the review thread.

> Yes, the first DOI may be wrong, but the system did not catch that.

While whedon's superpowers are very helpful, the system isn't perfect, and it's the job of the editor & @openjournals/joss-eics to give a human review. Clearly in this case a minor fix is needed.
