### 2019-03-09 838c9b4
adding more text to paper
Adding edits to meet JOSS requirements. Still need to proofread and add in references
