+++ paper.bib
@@ -1,3 +1,15 @@
+@article{Bendall,
+  title={Single-cell mass cytometry of differential immune and drug responses across a human hematopoietic continuum},
+  author={Bendall SC, Simonds EF, Qiu P, Amir el-AD, Krutzik PO, Finck R, Bruggner RV, Melamed R, Trejo A, Ornatsky OI, Balderas RS, Plevritis SK, Sachs K, Pe'er D, Tanner SD, Nolan GP.},
+  journal={Science},
+  volume={332},
+  number={6030},
+  pages={687-696},
+  year={2011},
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/science.1198704}
+}
+
 @article{Cytobank,
   title={Web-based Analysis and Publication of Flow Cytometry Experiments},
   author={Kotecha N, Krutzik PO, Irish JM.},
@@ -7,7 +19,7 @@
   pages={17:10.17.1–-10.17.24},
   year={2010},
   publisher={Wiley Online Library},
-  doi={https://doi.org/10.1002/0471142956.cy1017s53}
+  doi={10.1002/0471142956.cy1017s53}
 }
 
 @Manual{CytobankAPI,
@@ -16,4 +28,28 @@
   year = {2017},
   note = {R package version 1.1.0},
   url = {https://cran.r-project.org/web/packages/CytobankAPI/},
+}
+
+@article{polyarticularJIA,
+  title={Identification of enhanced IFN-γ signaling in polyarticular juvenile idiopathic arthritis with mass cytometry},
+  author={Throm AA, Moncrieffe H, Orandi AB, Pingel JT, Geurs TL, Miller HL, Daugherty AL, Malkova OM, Lovell DJ, Thompson SD, Grom AA, Cooper MA, Oh ST, French AR.},
+  journal={JCI Insight},
+  volume={3},
+  number={15},
+  pages={e121544},
+  year={2018},
+  publisher={American Society for Clinical Investigation},
+  doi={10.1172/jci.insight.121544}
+}
+
+@article{JDM,
+  title={Identification of enhanced IFN-γ signaling in polyarticular juvenile idiopathic arthritis with mass cytometry},
+  author={Throm AA, Alinger JB, Pingel JT,  Daugherty AL, Pachman LM, French AR.},
+  journal={JCI Insight}
+  volume={3},
+  number={22},
+  pages={e123236},
+  year={2018},
+  publisher={American Society for Clinical Investigation},
+  doi={10.1172/jci.insight.123236}
 }
