+++ paper.bib
@@ -45,7 +45,7 @@
 @article{JDM,
   title={Identification of enhanced IFN-γ signaling in polyarticular juvenile idiopathic arthritis with mass cytometry},
   author={Throm AA, Alinger JB, Pingel JT,  Daugherty AL, Pachman LM, French AR.},
-  journal={JCI Insight}
+  journal={JCI Insight},
   volume={3},
   number={22},
   pages={e123236},
