### 2019-03-11 #445 @athrom
@kyleniemeyer @pjotrp
I have updated the PDF to include more details as outlined in paper requirements and updated the doi with the version in Pubmed. Is there anything else I need to do to move the process forward?
