### 2019-03-12 482b3ea
Fixed some entries in paper.bib
Author names need to be separated with `and`, and curly brackets put around things in titles to retain capitalization. Also, the JDM paper didn't have the correct title (looked the same as polyarticularJIA)
