### 2019-03-12 dd676c2
Minor edits to paper.md
Fixed one citation, and added details to affiliations
