### 2019-03-12 #445 @kyleniemeyer
Hi @athrom, thanks for making those edits! The text looks great now.

I did notice some minor issues with the references, so I went ahead and fixed them (along with a few things in the paper) and submitted as two pull requests: https://github.com/athrom/CytobankAPIstats/pull/4 and https://github.com/athrom/CytobankAPIstats/pull/5

Could you merge those and let me know when?
