### 2019-03-13 #445 @athrom
@kyleniemeyer Thanks for helping with reference formatting. I have merged the pull requests.
