### 2019-03-13 #445 @kyleniemeyer
@athrom your paper is now published in JOSS!

Thanks to @nelkibs for reviewing and @pjotrp for editing!
